platform: thead: hardcode mbox for th1520
The mailbox binding for it isn't properly defined.

Hardcode it instead of requireing a non-standard DT node, and remove the
non-upstream "mbox" code from Xuantie (which is incompatible with later
upstream mailbox code).

The proper AON dt compatible string "thead,th1520-aon" is also added,
and a bogus ipc.h header file is cleaned up.
